projects/2-llm-rag/app/requirements.txt
Flask==3.0.2
chromadb-client==0.4.25.dev0
langchain==0.1.14
langchain-community==0.0.31
ollama==0.1.8
unstructured==0.11.8
